Fix SSIM data typing

+++ generative/metrics/ms_ssim.py
@@ -56,7 +56,7 @@ class MultiScaleSSIMMetric(RegressionMetric):
         data_range: float = 1.0,
         kernel_type: KernelType | str = KernelType.GAUSSIAN,
         kernel_size: int | Sequence[int, ...] = 11,
-        kernel_sigma: int | Sequence[int, ...] = 1.5,
+        kernel_sigma: float | Sequence[float, ...] = 1.5,
         k1: float = 0.01,
         k2: float = 0.03,
         weights: Sequence[float, ...] = (0.0448, 0.2856, 0.3001, 0.2363, 0.1333),
